deployment/requirements.txt
streamlit
pandas
numpy
tensorflow-cpu
Pillow